Updated the gitignore to include SEED questions.py is now creating an instance. And works on python2 Question tests now work correctly Test Scores file created for Dahar to finish

+++ questions.py
@@ -18,9 +18,9 @@ class Question:
   def get_correct_answer(self):
     return self.answer
 
-import csv
+import csv,os.path
 class QuestionImporter:
-  def import_questions(file='questions.csv'):
+  def import_questions(self, file='questions.csv'):
     reader = csv.reader(open(file), delimiter='|')
     questions = []
     for question in reader:
@@ -30,13 +30,11 @@ class QuestionImporter:
                 'a4' : question[5]}))
     return questions
 
-  def new_seed():
+  def new_seed(self):
     seed = 0
-    try:
-      seed_file = open('SEED','r')
-      seed_file.close()
-    except FileNotFoundError:
+    if not os.path.isfile('SEED'):
       seed_file = open('SEED','w')
+      print(str(seed))
       seed_file.write(str(seed))
       seed_file.close()
     seed_file = open('SEED','r') #open
@@ -48,7 +46,7 @@ class QuestionImporter:
     seed_file.close()
     return seed
 
-  def import_random_questions(amount=15):
+  def import_random_questions(self, amount=15):
     questions = import_questions()
     q_list = []
     seed = new_seed()
